+++ sws/loadtest.py
@@ -1,0 +1,1 @@
+"""A module to help quickly setup loadtesting for an application/service"""