+++ tests/framework_test.py
@@ -1,0 +1,7 @@
+"""Testing the functionality of sws.framework"""
+
+from sws.framework import *
+
+
+def test_stuff():
+    ...