+++ tests/loadtest_test .py
@@ -1,0 +1,7 @@
+"""Testing the functionality of sws.loadtest"""
+
+from sws.loadtest import *
+
+
+def test_stuff():
+    ...