+++ CHANGELOG.md
@@ -9,9 +9,10 @@ Features:
 - Moved ```trace``` command code to ```trace``` module and simplified tracing code
 - Added Bash autocomplete for ease of use on *nix systems
 - Added ```domains``` module and command
+- Added ```dns``` module and command
 - Added ```downloads`` module to make handling downlaods simpler
 - Added ```loadtest``` module and command
-- - Added ```framework``` module and command
+- Added ```framework``` module and command
 
 Documentation Improvements:
 - Added additional source, docs, and roadmap links to ```setup.py``` for PyPi